refactor: let Bazel shard live tests
