+++ e2e/BUILD.bazel
@@ -39,54 +39,6 @@ js_test(
     tags = ["manual"],
 )
 
-js_test(
-    name = "live_fast",
-    data = [
-        ":typecheck",
-        "//:node_modules/@perplexity-ai/perplexity_ai",
-        "//:node_modules/published-sdk",
-    ],
-    entry_point = "live.spec.ts",
-    env = {"LIVE_TEST_SHARD": "fast"},
-    tags = ["manual"],
-)
-
-js_test(
-    name = "live_responses",
-    data = [
-        ":typecheck",
-        "//:node_modules/@perplexity-ai/perplexity_ai",
-        "//:node_modules/published-sdk",
-    ],
-    entry_point = "live.spec.ts",
-    env = {"LIVE_TEST_SHARD": "responses"},
-    tags = ["manual"],
-)
-
-js_test(
-    name = "live_streaming_responses",
-    data = [
-        ":typecheck",
-        "//:node_modules/@perplexity-ai/perplexity_ai",
-        "//:node_modules/published-sdk",
-    ],
-    entry_point = "live.spec.ts",
-    env = {"LIVE_TEST_SHARD": "streaming-responses"},
-    tags = ["manual"],
-)
-
-js_test(
-    name = "live_async",
-    data = [
-        ":typecheck",
-        "//:node_modules/@perplexity-ai/perplexity_ai",
-        "//:node_modules/published-sdk",
-    ],
-    entry_point = "live.spec.ts",
-    env = {"LIVE_TEST_SHARD": "async"},
-    tags = ["manual"],
-)
-
 ts_project(
     name = "typecheck",
     testonly = True,
@@ -111,10 +63,7 @@ ts_project(
 
 ts_test(
     name = "e2e_test",
-    srcs = [
-        "live.spec.ts",
-        "sdk-parity.spec.ts",
-    ],
+    srcs = ["sdk-parity.spec.ts"],
     tags = ["manual"],
     tsconfig_types = ["node"],
     deps = [
