+++ e2e/live/helpers.ts
@@ -1,5 +1,4 @@
 import assert from 'node:assert/strict';
-import { describe, it } from 'node:test';
 import { setTimeout as delay } from 'node:timers/promises';
 // oxlint-disable-next-line no-restricted-imports -- Verify packaged public entrypoint.
 import { Perplexity as Candidate } from '@perplexity-ai/perplexity_ai';
@@ -29,7 +28,7 @@ interface SearchResponse {
   [key: string]: unknown;
 }
 
-function completionContract(response: unknown) {
+export function completionContract(response: unknown) {
   assert.ok(typeof response === 'object' && response !== null);
   const completion = response as CompletionResponse;
 
@@ -51,7 +50,7 @@ function completionContract(response: unknown) {
   };
 }
 
-function searchContract(response: unknown) {
+export function searchContract(response: unknown) {
   assert.ok(typeof response === 'object' && response !== null);
   const search = response as SearchResponse;
   assert.equal(typeof search.id, 'string');
@@ -74,7 +73,7 @@ function searchContract(response: unknown) {
   };
 }
 
-function embeddingContract(response: unknown) {
+export function embeddingContract(response: unknown) {
   assert.ok(typeof response === 'object' && response !== null);
   const embeddings = response as {
     data?: Array<{ embedding?: string; index?: number }>;
@@ -96,7 +95,7 @@ function embeddingContract(response: unknown) {
   };
 }
 
-function contextualizedEmbeddingContract(response: unknown) {
+export function contextualizedEmbeddingContract(response: unknown) {
   assert.ok(typeof response === 'object' && response !== null);
   const embeddings = response as {
     data?: Array<{ data?: Array<{ embedding?: string; index?: number }>; index?: number }>;
@@ -122,7 +121,7 @@ function contextualizedEmbeddingContract(response: unknown) {
   };
 }
 
-function responseContract(response: unknown) {
+export function responseContract(response: unknown) {
   assert.ok(typeof response === 'object' && response !== null);
   const result = response as {
     id?: string;
@@ -143,7 +142,7 @@ function responseContract(response: unknown) {
   };
 }
 
-function streamingCompletionContract(chunks: unknown[]) {
+export function streamingCompletionContract(chunks: unknown[]) {
   assert.ok(chunks.length > 0);
   const chunk = chunks.find((item) => {
     if (typeof item !== 'object' || item === null) {
@@ -166,7 +165,7 @@ function streamingCompletionContract(chunks: unknown[]) {
   };
 }
 
-function streamingResponseContract(events: unknown[]) {
+export function streamingResponseContract(events: unknown[]) {
   assert.ok(events.length > 0);
   const types = events.map((event) => {
     assert.ok(typeof event === 'object' && event !== null);
@@ -186,7 +185,7 @@ function streamingResponseContract(events: unknown[]) {
   };
 }
 
-function asyncCompletionContract(response: unknown) {
+export function asyncCompletionContract(response: unknown) {
   assert.ok(typeof response === 'object' && response !== null);
   const completion = response as {
     created_at?: unknown;
@@ -221,7 +220,7 @@ function asyncCompletionStatus(response: unknown) {
   return status;
 }
 
-async function pollAsyncCompletion(initial: unknown, retrieve: (id: string) => PromiseLike<unknown>) {
+export async function pollAsyncCompletion(initial: unknown, retrieve: (id: string) => PromiseLike<unknown>) {
   const id = asyncCompletionID(initial);
   let completion = initial;
 
@@ -238,7 +237,7 @@ async function pollAsyncCompletion(initial: unknown, retrieve: (id: string) => P
   assert.fail(`Async completion ${id} did not finish`);
 }
 
-async function collect(stream: AsyncIterable<unknown>) {
+export async function collect(stream: AsyncIterable<unknown>) {
   const chunks: unknown[] = [];
   for await (const chunk of stream) {
     chunks.push(chunk);
@@ -246,7 +245,7 @@ async function collect(stream: AsyncIterable<unknown>) {
   return chunks;
 }
 
-async function expectMatchingContracts(
+export async function expectMatchingContracts(
   candidateCall: () => PromiseLike<unknown>,
   publishedCall: () => PromiseLike<unknown>,
   contract: (response: unknown) => unknown,
@@ -255,172 +254,11 @@ async function expectMatchingContracts(
   assert.deepStrictEqual(contract(candidateResponse), contract(publishedResponse));
 }
 
-function createClients() {
+export function createClients() {
   const apiKey = process.env['PPLX_API_TOKEN'];
   assert.ok(apiKey, 'PPLX_API_TOKEN must be set');
   return {
     candidate: new Candidate({ apiKey, maxRetries: 0 }),
     published: new Published({ apiKey, maxRetries: 0 }),
   };
-}
-
-const liveTestShard = process.env['LIVE_TEST_SHARD'];
-const runsShard = (shard: string) => !liveTestShard || liveTestShard === shard;
-
-if (runsShard('fast')) {
-  describe('live API parity: fast APIs', () => {
-    it('matches published chat completion response contract', async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        max_tokens: 16,
-        messages: [{ content: 'Reply with only the word pong.', role: 'user' as const }],
-        model: 'sonar',
-        temperature: 0,
-      };
-
-      await expectMatchingContracts(
-        () => candidate.chat.completions.create(request),
-        () => published.chat.completions.create(request),
-        completionContract,
-      );
-    });
-
-    it('matches published streaming chat contract', async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        max_tokens: 16,
-        messages: [{ content: 'Reply with only the word pong.', role: 'user' as const }],
-        model: 'sonar',
-        stream: true as const,
-        temperature: 0,
-      };
-
-      await expectMatchingContracts(
-        async () => collect(await candidate.chat.completions.create(request)),
-        async () => collect(await published.chat.completions.create(request)),
-        (chunks) => {
-          assert.ok(Array.isArray(chunks));
-          return streamingCompletionContract(chunks);
-        },
-      );
-    });
-
-    it('matches published search response contract', async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        max_results: 1,
-        query: 'Perplexity AI',
-      };
-
-      await expectMatchingContracts(
-        () => candidate.search.create(request),
-        () => published.search.create(request),
-        searchContract,
-      );
-    });
-
-    it('matches published embeddings response contract', async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        dimensions: 128,
-        input: 'Perplexity answers questions.',
-        model: 'pplx-embed-v1-0.6b' as const,
-      };
-
-      await expectMatchingContracts(
-        () => candidate.embeddings.create(request),
-        () => published.embeddings.create(request),
-        embeddingContract,
-      );
-    });
-
-    it('matches published contextualized embeddings response contract', async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        dimensions: 128,
-        input: [['Perplexity answers questions.', 'Its answers include citations.']],
-        model: 'pplx-embed-context-v1-0.6b' as const,
-      };
-
-      await expectMatchingContracts(
-        () => candidate.contextualizedEmbeddings.create(request),
-        () => published.contextualizedEmbeddings.create(request),
-        contextualizedEmbeddingContract,
-      );
-    });
-  });
-}
-
-if (runsShard('responses')) {
-  describe('live API parity: Responses API', () => {
-    it('matches published responses API contract', async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        input: 'Reply with only the word pong.',
-        max_output_tokens: 16,
-        preset: 'pro-search',
-      };
-
-      await expectMatchingContracts(
-        () => candidate.responses.create(request),
-        () => published.responses.create(request),
-        responseContract,
-      );
-    });
-  });
-}
-
-if (runsShard('streaming-responses')) {
-  describe('live API parity: streaming Responses API', () => {
-    it('matches published streaming responses API contract', async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        input: 'Reply with only the word pong.',
-        max_output_tokens: 16,
-        preset: 'pro-search',
-        stream: true as const,
-      };
-
-      await expectMatchingContracts(
-        async () => collect(await candidate.responses.create(request)),
-        async () => collect(await published.responses.create(request)),
-        (events) => {
-          assert.ok(Array.isArray(events));
-          return streamingResponseContract(events);
-        },
-      );
-    });
-  });
-}
-
-if (runsShard('async')) {
-  describe('live API parity: async API', () => {
-    it('matches published async chat completion lifecycle', { timeout: 180_000 }, async () => {
-      const { candidate, published } = createClients();
-      const request = {
-        request: {
-          max_tokens: 16,
-          messages: [{ content: 'Reply with only the word pong.', role: 'user' as const }],
-          model: 'sonar-deep-research',
-        },
-      };
-      const [candidateInitial, publishedInitial] = await Promise.all([
-        candidate.async.chat.completions.create(request),
-        published.async.chat.completions.create(request),
-      ]);
-      assert.deepStrictEqual(
-        asyncCompletionContract(candidateInitial),
-        asyncCompletionContract(publishedInitial),
-      );
-
-      const [candidateCompleted, publishedCompleted] = await Promise.all([
-        pollAsyncCompletion(candidateInitial, (id) => candidate.async.chat.completions.get(id)),
-        pollAsyncCompletion(publishedInitial, (id) => published.async.chat.completions.get(id)),
-      ]);
-      assert.deepStrictEqual(
-        asyncCompletionContract(candidateCompleted),
-        asyncCompletionContract(publishedCompleted),
-      );
-    });
-  });
 }
